+++ packages/whisky-wallet/tests/wallet/mod.rs
@@ -1,3 +1,3 @@
 mod address;
 mod bip39;
-mod fetcher;
+mod provider;
